make-entities (#10)
* refactor services

* config hibernate

* move memoryImpl

* fix Handlers

* add HibernateConfiguration

* fix Listeners

* remove deprecated services

* add new services

* fix entities

* add UserRole

+++ src/main/java/ru/shummi/operation/impl/AccountCloseHandler.java
@@ -28,7 +28,7 @@ public class AccountCloseHandler implements OperationHandler {
 
     @Override
     public String perform() {
-        int accountId = ioHandler.getInteger("Enter account ID to close");
+        long accountId = ioHandler.getLong("Enter account ID to close");
         userAccountService.closeAccountById(accountId);
 
         return "Account with ID %s has been closed.".formatted(accountId);
